pyproject.toml
[project]
name = "purz"
description = ""
version = "1.7.0"
license = {file = "LICENSE"} 
# classifiers = [
#     # For OS-independent nodes (works on all operating systems)
#     "Operating System :: OS Independent",
# 
#     # OR for OS-specific nodes, specify the supported systems:
#     "Operating System :: Microsoft :: Windows",  # Windows specific
#     "Operating System :: POSIX :: Linux",  # Linux specific
#     "Operating System :: MacOS",  # macOS specific
#     
#     # GPU Accelerator support. Pick the ones that are supported by your extension.
#     "Environment :: GPU :: NVIDIA CUDA",    # NVIDIA CUDA support
#     "Environment :: GPU :: AMD ROCm",       # AMD ROCm support
#     "Environment :: GPU :: Intel Arc",      # Intel Arc support
#     "Environment :: NPU :: Huawei Ascend",  # Huawei Ascend support
#     "Environment :: GPU :: Apple Metal",    # Apple Metal support
# ]

dependencies = ["opencv-python>=4.5.0", "Pillow>=8.0.0", "torch>=1.9.0", "numpy>=1.19.0"]

[project.urls]
Repository = "https://github.com/purzbeats/ComfyUI-Purz"
#  Used by Comfy Registry https://registry.comfy.org
Documentation = "https://github.com/purzbeats/ComfyUI-Purz/wiki"
"Bug Tracker" = "https://github.com/purzbeats/ComfyUI-Purz/issues"

[tool.comfy]
PublisherId = "purz"
DisplayName = "ComfyUI-Purz"
Icon = ""
includes = [] 
# "requires-comfyui" = ">=1.0.0"  # ComfyUI version compatibility

